requirements.txt
# Core dependencies for AI continuation of Dream of the Red Chamber
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langchain-openai==0.0.2
openai==1.6.1

# Text processing and analysis
jieba==0.42.1
pypinyin==0.50.0
regex==2023.12.25

# Data handling
pandas==2.1.4
numpy==1.26.2
pydantic==2.5.2

# Configuration and environment
python-dotenv==1.0.0
PyYAML==6.0.1

# CLI and user interface
click==8.1.7
rich==13.7.0
tqdm==4.66.1

# Logging and monitoring
loguru==0.7.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0

# Optional: For advanced features
# transformers==4.36.2  # If you want to use local models
# torch==2.1.2  # For local model support
# faiss-cpu==1.7.4  # For vector search if needed